requirements.txt
streamlit
matplotlib
pystac-client
planetary-computer
geopandas
rioxarray
leafmap
sqlalchemy
azure-storage-blob
azure-identity
numpy
